More WIP, trying JLH score instead of NGD
Unfortunately, JLH results in too many common words making the top of the list. The stemming approach being used for English may also be a problem; see the TODO note where the stopWordStems bit is commented out. The idea of filtering out common words first and then simply sorting by frequency may not be a bad one, but we're going to need to rethink how we're doing this regardless...